[Endpoint] Added Post for product creation

+++ StoreBackend.Api/DTOs/CreateProductDTO.cs
@@ -6,6 +6,6 @@ public record class CreateProductDTO
 (
     [Required][StringLength(20)] string name,
     [Required] int initialStock,
-    [Required] decimal price,
+    [Required] double price,
     [Required] int productType
 );
